Another warning treated as error: int to double. Use static_cast for both

+++ src/energyplus/ForwardTranslator/ForwardTranslateScheduleFixedInterval.cpp
@@ -214,7 +214,7 @@ boost::optional<IdfObject> ForwardTranslator::translateScheduleFixedInterval( Sc
         // Write out the current entry
         Time time(0,0,0,secondsFromStartOfDay);
         int hours = time.hours();
-        int minutes = time.minutes() + floor((time.seconds()/60.0) + 0.5);
+        int minutes = time.minutes() + static_cast<int>(std::floor((time.seconds()/60.0) + 0.5));
         // This is a little dangerous, but all of the problematic 24:00
         // times that might need to cause a day++ should be caught above.
         if(minutes==60)
